Na tela 1, deve adicionar o ícone coração na lista de cards All e Favoritos. A ideia é marcar o card como favorito ou não a. O ícone de coração deve ficar em cima da imagem da planta; b. Ao clicar clicar no icone de coração o card da planta deve ser removida da lista all, e ser mostrada na lista de favoritos; c. Ao clicar no icone de coração na lista de favoritos, que vai está marcado, o card da planta deve ser removido da lista de favorito e ser mostrado na lista All

+++ Intermediario/desafio/lib/screens/detail/components/plant_detail_body.dart
@@ -41,9 +41,7 @@ class PlantDetailBody extends StatelessWidget {
         const Spacer(),
         const Spacer(),
         TitleAndPrice(
-            title: plant.name,
-            country: plant.country,
-            price: plant.price.toInt()),
+            title: plant.name, country: plant.country, price: plant.price),
         const Spacer(),
         Row(
           children: [
